Split CLI into subcommands with per-command modules
- Each action (download, check, clean-cache, show-auth-script) is a separate subcommand in its own file under `cli/commands/`
- `main.py` is now a thin shell: app setup, command registration, and error handling
- `clean-cache` works directly with SQLitePostCache, no HTTP session or AppEnvironment needed
- App initialization logic (config, update check, AppEnvironment) extracted to `application/di/initialized_app.py`
